+++ src/hooks/useTelegramAuth.js
@@ -1,0 +1,40 @@
+import { useEffect, useState } from 'react';
+import { init, initData } from '@telegram-apps/sdk';
+
+// Хук для проверки авторизации через Telegram
+export const useTelegramAuth = () => {
+    const [isAuthorized, setIsAuthorized] = useState(null); // null = загрузка
+
+    useEffect(() => {
+        console.log(initData, "Init Data 123")
+        const checkAuth = async () => {
+            try {
+                await init(); // Инициализируем SDK
+
+                // Проверяем наличие initData
+                if (!initData.exists()) {
+                    setIsAuthorized(false);
+                    return;
+                }
+
+                // Проверка подписи (реализуйте на бэкенде)
+                const isValid = await validateInitData(initData.raw());
+                setIsAuthorized(isValid);
+            } catch (error) {
+                console.error('Ошибка проверки авторизации:', error);
+                setIsAuthorized(false);
+            }
+        };
+
+        checkAuth();
+    }, []);
+
+    return isAuthorized;
+};
+
+// Заглушка для валидации подписи (реализуйте на бэкенде)
+const validateInitData = async (rawInitData) => {
+    // TODO: Отправьте rawInitData на сервер для проверки
+    // Возвращайте true, если подпись валидна
+    return true;
+};
